requirements.txt
aiohttp==3.4.4
async-timeout==3.0.0
attrs==18.2.0
aws-lambda-logging==0.1.1
boto3==1.9.14
botocore==1.12.14
chardet==3.0.4
docutils==0.14
idna==2.7
jmespath==0.9.3
multidict==4.4.2
python-dateutil==2.7.3
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
slackweb==1.0.5
urllib3==1.23
yarl==1.2.6